add intermediate Slice component and rename Slice to Table
This way, Scroller is more stable, it should not change often.
